requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
python-telegram-bot==9.0.0
emoji==0.4.5
